Changes in button assignments to USERFUNCx  fixed
Changes on the Buttons tab of ‘PiJuice HAT Configuration’ are updated
in the PiJuice MCU when you press the ‘Apply’ button.
However if you change e.g. USERFUNC6 to USERFUNC7 for a button press,
this change does not get picked up by the pijuice service.
It will still execute USERFUNC6 after the change.
This is because the USERFUNCx assignments are only picked up when the
service starts.

Easy fix: Include the USERFUNCx assignments in reload_settings() of
pijuice_sys.py and add a notify_service() call in pijuice_gui.py when
you press the ‘Apply’ button