Bug fixes for updating status and UTC time
Updating the status is done by a recurring alarm. However alarm was not stopped when leaving the status tab.
Same for the UTC update in the wakeupalarm tab.
This caused the crash reported in issue #163.
In addition the check for unchanged LED settings was removed. The firmware sets D2 to bright blue after firmware update and to bright green after a parameter reset, but the D2 user settings are not updated accordingly.