Reload rtc module when original rtc probe fails
When RPi is off and PiJuice in low power mode then after applying power to the RPI the PiJuice changes from low power to run mode too late to respond to the i2c rtc probe on address 0x68.
To remedy this the pijuice service now checks for this and if needed reloads the rtc_ds1307 module which makes the rtc available to the os (-> sudo hwclock -r works again).
In addition the pijuice service is started when the i2c device is available instead of waiting for the network.