Wake up the clock before loading rtc_ds1307 module
Under Fedora 39 with kernel 6.5.6, loading the module rtc_ds1307 does
not result in a functioning RTC. The clock needs to be read (woken up?)
before the module is loaded. This change adds a read of the clock with a
timeout.

In addition, the /dev/rtc device file takes a fraction of a second to
load so the test for its existence alway fails. A 1 second sleep gives
time for the file to be created.

It's not clear to me if eventually this will be required for Raspberry
OS when it eventually gets to the 6.5.6 kernel. However it seems this
change should always work on kernel 5.

Also fix a misspelling of module in the success message.