Set RTC to system time if it was reset to 0
When the RTC is reset to 0, for example from a loss of battery, it wreaks
havoc on the systemd log and probably other subsystems. systemd will set
the system time based on a recent file timestamp in the filesystem (how it
does this depends if chrony or systemd.timesync is being used). But then when
rtc_ds1307 loads, the kernel module sets the system time to the RTC's Jan 1,
2000. If network time is running it will eventually figure it out but the
logs are messed up with the huge jumps in time.

The best place to address this is just before the modprobe of
rtc_ds1307. If the year is 2000, then most certainly the RTC has been
reset. The convenient SetTime() method sets it to the current system
time which is most certainly much closer to the correct time than
year 2000.