Add Firmware DT instructions
It turns out kernel Device Tree is a bit different from firmware Device
Tree. So we need to add instructions to switch Fedora over to firmware
DT and enable ds1307.

I didn't realize this previously because Fedora 39's kernel-core-6.5.6
has a bug in it. The dtb files don't get place into /boot which causes
Fedora to fall back to firmware DT. So this whole time I was in firmware
DT mode.